+++ rts/c/tuning.h
@@ -1,0 +1,40 @@
+// Start of tuning.h
+
+static char* load_tuning_file(const char *fname,
+                              void *cfg,
+                              int (*set_size)(void*, const char*, size_t)) {
+  const int max_line_len = 1024;
+  char* line = malloc(max_line_len);
+
+  FILE *f = fopen(fname, "r");
+
+  if (f == NULL) {
+    snprintf(line, max_line_len, "Cannot open file: %s", strerror(errno));
+    return line;
+  }
+
+  int lineno = 0;
+  while (fgets(line, max_line_len, f) != NULL) {
+    lineno++;
+    char *eql = strstr(line, "=");
+    if (eql) {
+      *eql = 0;
+      int value = atoi(eql+1);
+      if (set_size(cfg, line, value) != 0) {
+        strncpy(eql+1, line, max_line_len-strlen(line)-1);
+        snprintf(line, max_line_len, "Unknown name '%s' on line %d.", eql+1, lineno);
+        return line;
+      }
+    } else {
+      snprintf(line, max_line_len, "Invalid line %d (must be of form 'name=int').",
+               lineno);
+      return line;
+    }
+  }
+
+  free(line);
+
+  return NULL;
+}
+
+// End of tuning.h