+++ rts/python/tuning.py
@@ -1,0 +1,11 @@
+### start of tuning.py
+###
+### Reading the .tuning file.
+
+def read_tuning_file(kvs, f):
+    for line in f.read().splitlines():
+        size, value = line.split('=')
+        kvs[size] = int(value)
+    return kvs
+
+### end of tuning.py