+++ src/Language/Futhark/Pretty.hs
@@ -237,7 +237,7 @@ instance (Eq vn, IsName vn, Annot f) => Pretty (ExpBase f vn) where
      then equals </> indent 2 (ppr e)
      else equals <+> align (ppr e)) </>
     (case body of LetPat{} -> ppr body
-                  _        -> text "in" <+> ppr body)
+                  _        -> text "in" <+> align (ppr body))
     where linebreak = case e of
                         Map{}       -> True
                         Reduce{}    -> True
@@ -297,9 +297,6 @@ instance (Eq vn, IsName vn, Annot f) => Pretty (ExpBase f vn) where
   pprPrec _ (Scan lam e a _) = ppSOAC "scan" [lam] [e, a]
   pprPrec _ (Filter lam a _) = ppSOAC "filter" [lam] [a]
   pprPrec _ (Partition k lam a _) = text "partition" <+> ppr k <+> spread (map (pprPrec 10) [lam, a])
-  pprPrec _ (Zip 0 e es _ _) = text "zip" <+> spread (map (pprPrec 10) (e:es))
-  pprPrec _ (Zip i e es _ _) = text "zip@" <> ppr i <+> spread (map (pprPrec 10) (e:es))
-  pprPrec _ (Unzip e _ _) = text "unzip" <+> pprPrec (-1) e
   pprPrec _ (Unsafe e _) = text "unsafe" <+> pprPrec (-1) e
   pprPrec _ (Assert e1 e2 _ _) = text "assert" <+> pprPrec 10 e1 <+> pprPrec 10 e2
   pprPrec p (Lambda tparams params body ascript _ _) =